channelz: refactor to move proto API to a separate package
