+++ channelz/internal/protoconv/sockopt_linux.go
@@ -16,7 +16,7 @@
  *
  */
 
-package service
+package protoconv
 
 import (
 	"time"
